changed real(8) to real(real64) and integer to integer(int32)

+++ src/ControlParameterClass/ControlParameterClass.f90
@@ -1,10 +1,11 @@
 module ControlParaeterClass
+    use, intrinsic :: iso_fortran_env    
     implicit none
     type::ControlParameter_
-        real(8) :: Tol
-        integer :: SimMode
-        integer :: ItrTol
-        integer :: Timestep
+        real(real64) :: Tol
+        integer(int32) :: SimMode
+        integer(int32) :: ItrTol
+        integer(int32) :: Timestep
     contains
         procedure,public :: SetControlPara => SetControlPara
     end type ControlParameter_
@@ -14,8 +15,8 @@ contains
 !######### Import Control Parameters #########
 subroutine SetControlPara(obj,OptionalTol,OptionalItrTol,OptionalTimestep,OptionalSimMode)
     class(ControlParameter_),intent(out)::obj
-    real(8),optional,intent(in)::OptionalTol
-    integer,optional,intent(in)::OptionalSimMode,OptionalItrTol,OptionalTimestep
+    real(real64),optional,intent(in)::OptionalTol
+    integer(int32),optional,intent(in)::OptionalSimMode,OptionalItrTol,OptionalTimestep
 
     if(present(OptionalSimMode) )then
         obj%SimMode = OptionalSimMode
