+++ src/OpenMPClass/OpenMPClass.f90
@@ -1,12 +1,13 @@
 module OpenMPClass
+    use, intrinsic :: iso_fortran_env
     use omp_lib
     implicit none
 contains
 
 !#########################################################
 subroutine omp_dot_product1(a,b,sum)
-    real(8),intent(in)::a(:),b(:)
-    real(8),intent(out) ::sum
+    real(real64),intent(in)::a(:),b(:)
+    real(real64),intent(out) ::sum
     
 !$OMP parallel 
 !$omp workshare
@@ -19,10 +20,10 @@ end subroutine omp_dot_product1
 
 !#########################################################
 subroutine omp_dot_product(a,b,sum)
-    real(8),intent(in)::a(:),b(:)
-    real(8),intent(out) ::sum
-    integer :: i
-    real(8) :: psum
+    real(real64),intent(in)::a(:),b(:)
+    real(real64),intent(out) ::sum
+    integer(int32) :: i
+    real(real64) :: psum
 sum=0
 !$OMP parallel private(i,psum)
 !$omp do reduction(+: sum)
@@ -37,10 +38,10 @@ end subroutine omp_dot_product
 
 !#########################################################
 subroutine omp_dot_product2(a,b,sum)
-    real(8),intent(in)::a(:),b(:)
-    real(8),intent(out) ::sum
-    integer :: i
-    real(8) :: psum
+    real(real64),intent(in)::a(:),b(:)
+    real(real64),intent(out) ::sum
+    integer(int32) :: i
+    real(real64) :: psum
 sum=0
 !$OMP parallel
 !$omp do
@@ -54,9 +55,9 @@ end subroutine omp_dot_product2
 
 !#########################################################
 subroutine omp_matmul(a,b,mm)
-    real(8),intent(in)::a(:,:),b(:)
-    real(8),intent(inout)::mm(:)
-    integer :: i
+    real(real64),intent(in)::a(:,:),b(:)
+    real(real64),intent(inout)::mm(:)
+    integer(int32) :: i
 
 do i=1,size(a,1)
     !$OMP parallel
